+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.proton` artifacts to version `0.54`. This change adds basic profiling support that can be enabling by setting the annotation processor option `sting.profile` to `true`.
+
 ### [v0.22](https://github.com/sting-ioc/sting/tree/v0.22) (2021-11-26) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.21...v0.22)
 
 Changes in this release: